requirements.txt
aiohttp==3.9.5
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
contourpy==1.3.2
cycler==0.12.1
diskcache==5.6.3
distro==1.9.0
filelock==3.18.0
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.5.1
gguf==0.17.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.23.5
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.10.0
kiwisolver==1.4.8
llama_cpp_python==0.3.9
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.5.0
networkx==3.5
numpy==1.26.4
openai==1.55.3
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pluggy==1.6.0
prometheus_client==0.20.0
propcache==0.3.2
protobuf==4.25.8
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
pytest==8.3.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
safetensors==0.5.3
seaborn==0.13.2
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tokenizers==0.20.3
torch==2.2.2
tqdm==4.67.1
transformers==4.46.3
typer==0.15.4
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
wget==3.2
yarl==1.20.1
